fix: not a list objectives, int argument
- the order in the ArgumentResponse value types, Union[int, float,...]
  caused float values to be int'ified, -> Union[float, int, ...], fixed
  the issue
- to get objectives dropdown list, it used to return the keys of a
  dictionnary, but was not considered as a list, so now explicit
  conversion to list

+++ api/bioptim_gui_api/acrobatics_ocp/endpoints/acrobatics_responses.py
@@ -90,7 +90,7 @@ class WeightResponse(WeightRequest):
 class ArgumentResponse(BaseModel):
     key: str
     type: str
-    value: Union[int, float, str, list, None, Axis]
+    value: Union[float, int, str, list, None, Axis]
 
 
 class NewGeneratedBioMod(NamedTuple):
